+++ src/main/java/com/hubspot/httpql/impl/QueryParser.java
@@ -7,7 +7,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.ServiceLoader;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -15,9 +14,6 @@ import java.util.stream.Collectors;
 import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.core.UriInfo;
 
-import org.apache.commons.lang.BooleanUtils;
-import org.apache.commons.lang.math.NumberUtils;
-import org.jooq.impl.DSL;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -26,17 +22,13 @@ import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import com.fasterxml.jackson.databind.introspect.BeanPropertyDefinition;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.CaseFormat;
-import com.google.common.base.Splitter;
 import com.google.common.base.Strings;
 import com.google.common.collect.HashBasedTable;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Table;
-import com.hubspot.httpql.ConditionProvider;
 import com.hubspot.httpql.DefaultMetaUtils;
-import com.hubspot.httpql.Filter;
 import com.hubspot.httpql.MetaQuerySpec;
-import com.hubspot.httpql.MultiParamConditionProvider;
 import com.hubspot.httpql.ParsedQuery;
 import com.hubspot.httpql.QueryConstraints;
 import com.hubspot.httpql.QuerySpec;
@@ -46,7 +38,6 @@ import com.hubspot.httpql.error.ConstraintType;
 import com.hubspot.httpql.error.ConstraintViolation;
 import com.hubspot.httpql.error.FilterViolation;
 import com.hubspot.httpql.error.LimitViolationType;
-import com.hubspot.httpql.filter.Equal;
 import com.hubspot.httpql.internal.BoundFilterEntry;
 import com.hubspot.httpql.internal.FilterEntry;
 import com.hubspot.httpql.internal.MultiValuedBoundFilterEntry;
@@ -60,44 +51,31 @@ import com.sun.jersey.core.util.MultivaluedMapImpl;
  * The parser's job is to take a set of query arguments (string key/value pairs) and turn it into a high-level query representation, assuming it is valid according to the defined filtering rules (provided via {@link FilterBy} and
  * {@link OrderBy} annotations).
  *
- * @author tdavis
  */
 public class QueryParser<T extends QuerySpec> {
   private static final Logger LOG = LoggerFactory.getLogger(QueryParser.class);
-  private static final ServiceLoader<Filter> LOADER = ServiceLoader.load(Filter.class);
-  private static final Map<String, Filter> BY_NAME = new HashMap<>();
-  private static final Set<String> RESERVED_WORDS = ImmutableSet.of("offset", "limit", "order", "includeDeleted");
-  private static final Set<String> IGNORED_PARAMS = ImmutableSet.of("hapikey", "access_token", "clienttimeout");
-
-  static {
-    for (Filter filter : LOADER) {
-      for (String name : filter.names()) {
-        BY_NAME.put(name, filter);
-      }
-    }
-  }
-
-  private static final Splitter FILTER_PARAM_SPLITTER = Splitter.on("__").trimResults();
-  private static final Splitter MULTIVALUE_PARAM_SPLITTER = Splitter.on(",").omitEmptyStrings().trimResults();
 
   private static final Function<String, String> SNAKE_CASE_TRANSFORMER = input -> CaseFormat.LOWER_CAMEL.to(CaseFormat.LOWER_UNDERSCORE, input);
+  private static final Set<String> RESERVED_WORDS = ImmutableSet.of("offset", "limit", "order", "includeDeleted");
 
   private final Class<T> queryType;
   protected final Collection<String> orderableFields;
   private final ObjectMapper mapper;
   private final MetaQuerySpec<T> meta;
-  private final boolean strictMode;
 
   private int maxLimit;
   private int maxOffset;
   private int defaultLimit = 100;
+  private final boolean strictMode;
+  private final UriParamParser uriParamParser;
 
-  protected QueryParser(final Class<T> spec, boolean strictMode, MetaQuerySpec<T> meta) {
+  protected QueryParser(final Class<T> spec, boolean strictMode, MetaQuerySpec<T> meta, UriParamParser uriParamParser) {
     this.orderableFields = new ArrayList<>();
     this.queryType = spec;
     this.mapper = Rosetta.getMapper();
     this.meta = meta;
     this.strictMode = strictMode;
+    this.uriParamParser = uriParamParser;
   }
 
   protected void setConstraints(final Class<T> spec) {
@@ -129,92 +107,19 @@ public class QueryParser<T extends QuerySpec> {
     return parse(uriInfo.getQueryParameters());
   }
 
-  public static MultivaluedMap<String, String> multimapToMultivaluedMap(Multimap<String, String> map) {
-
-    MultivaluedMap<String, String> result = new MultivaluedMapImpl();
-    for (Map.Entry<String, String> entry : map.entries()) {
-      result.add(entry.getKey(), entry.getValue());
-    }
-
-    return result;
-  }
-
   public ParsedQuery<T> parse(Multimap<String, String> uriParams) {
-    return parse(multimapToMultivaluedMap(uriParams));
+    return parse(uriParamParser.multimapToMultivaluedMap(uriParams));
   }
 
   public ParsedQuery<T> createEmptyQuery() {
     return parse(new MultivaluedMapImpl());
   }
 
-  public static ParsedUriParams parseUriParams(Multimap<String, String> uriParams) {
-    return parseUriParams(multimapToMultivaluedMap(uriParams));
-  }
-
-  public static ParsedUriParams parseUriParams(MultivaluedMap<String, String> uriParams) {
-
-    final ParsedUriParams result = new ParsedUriParams();
-
-    // make a copy so we can modify it
-    MultivaluedMap<String, String> params = new MultivaluedMapImpl();
-    for (Map.Entry<String, List<String>> entry : uriParams.entrySet()) {
-      if (!IGNORED_PARAMS.contains(entry.getKey().toLowerCase())) {
-        params.put(entry.getKey(), entry.getValue());
-      }
-    }
-
-    result.setIncludeDeleted(BooleanUtils.toBoolean(params.getFirst("includeDeleted")));
-    params.remove("includeDeleted");
-
-    final int limit = NumberUtils.toInt(params.getFirst("limit"), 0);
-    if (limit != 0) {
-      result.setLimit(limit);
-    }
-    params.remove("limit");
-
-    final int offset = NumberUtils.toInt(params.getFirst("offset"), 0);
-    if (offset != 0) {
-      result.setOffset(offset);
-    }
-    params.remove("offset");
-
-    result.addOrderBys(params.get("order"));
-    params.remove("order");
-    result.addOrderBys(params.get("orderBy"));
-    params.remove("orderBy");
-
-    for (Map.Entry<String, List<String>> entry : params.entrySet()) {
-      List<String> parts = FILTER_PARAM_SPLITTER.splitToList(entry.getKey().trim());
-      if (parts.size() > 2) {
-        continue;
-      }
-
-      final String fieldName = parts.get(0);
-      final String filterName = filterNameFromParts(parts);
-      final Filter filter = BY_NAME.get(filterName);
-
-      if (filter == null) {
-        throw new FilterViolation(String.format("Unknown filter type `%s`", filterName));
-      }
-
-      List<String> values = entry.getValue();
-      ConditionProvider conditionProvider = filter.getConditionProvider(DSL.field(fieldName));
-
-      if (conditionProvider instanceof MultiParamConditionProvider && values.size() == 1 && values.get(0).contains(",")) {
-        values = MULTIVALUE_PARAM_SPLITTER.splitToList(values.get(0));
-      }
-
-      result.addFieldFilter(new FieldFilter(filter, filterName, fieldName, values));
-    }
-
-    return result;
-  }
-
   public ParsedQuery<T> parse(MultivaluedMap<String, String> uriParams) {
     final Map<String, Object> fieldValues = new HashMap<>();
     final List<BoundFilterEntry<T>> boundFilterEntries = new ArrayList<>();
 
-    final ParsedUriParams parsedUriParams = parseUriParams(uriParams);
+    final ParsedUriParams parsedUriParams = uriParamParser.parseUriParams(uriParams);
 
     final boolean includeDeleted = parsedUriParams.isIncludeDeleted();
     final Optional<Integer> limit = getLimit(parsedUriParams.getLimit());
@@ -240,7 +145,7 @@ public class QueryParser<T extends QuerySpec> {
       }
 
       String finalFieldName = fieldName;
-      Optional<BoundFilterEntry<T>> filterEntryOptional = filterTable.rowKeySet().stream().filter(f -> Objects.equals(f.getFieldName(), finalFieldName) && Objects.equals(f.getFilter(), BY_NAME
+      Optional<BoundFilterEntry<T>> filterEntryOptional = filterTable.rowKeySet().stream().filter(f -> Objects.equals(f.getFieldName(), finalFieldName) && Objects.equals(f.getFilter(), UriParamParser.BY_NAME
           .get(filterName))).findFirst();
 
       // Use reserved words instead of simple look-up to throw exception on disallowed fields
@@ -263,7 +168,7 @@ public class QueryParser<T extends QuerySpec> {
 
       if (prop.getPrimaryMember() != null) {
         ann = prop.getPrimaryMember().getAnnotation(FilterBy.class);
-        if (Strings.emptyToNull(ann.as()) != null) {
+        if (ann != null && Strings.emptyToNull(ann.as()) != null) {
           boundColumn.setActualField(fieldMap.get(ann.as()));
         }
       }
@@ -365,14 +270,6 @@ public class QueryParser<T extends QuerySpec> {
     return orderings;
   }
 
-  private static String filterNameFromParts(List<String> parts) {
-    if (parts.size() == 1) {
-      return (new Equal()).names()[0];
-    } else {
-      return parts.get(1);
-    }
-  }
-
   public Class<T> getQueryType() {
     return queryType;
   }
@@ -391,6 +288,7 @@ public class QueryParser<T extends QuerySpec> {
     protected Class<T> spec;
     protected boolean strictMode;
     protected MetaQuerySpec<T> meta;
+    protected UriParamParser uriParamParser;
 
     public Builder(Class<T> spec) {
       this.spec = spec;
@@ -411,6 +309,11 @@ public class QueryParser<T extends QuerySpec> {
       return this;
     }
 
+    public Builder<T> withParamParser(UriParamParser uriParamParser) {
+      this.uriParamParser = uriParamParser;
+      return this;
+    }
+
     @SuppressWarnings("unchecked")
     public QueryParser<T> build() {
       QueryParser<T> qp = (QueryParser<T>) CACHED_PARSERS.get(spec, strictMode);
@@ -422,9 +325,13 @@ public class QueryParser<T extends QuerySpec> {
         meta = new DefaultMetaQuerySpec<>(spec);
       }
 
+      if (uriParamParser == null) {
+        uriParamParser = UriParamParser.newBuilder().build();
+      }
+
       Map<String, BeanPropertyDefinition> fields = BeanPropertyIntrospector.getFields(spec);
 
-      qp = new QueryParser<>(spec, strictMode, meta);
+      qp = new QueryParser<>(spec, strictMode, meta, uriParamParser);
       qp.setConstraints(spec);
       qp.buildOrderableFields(fields);
 
@@ -432,10 +339,6 @@ public class QueryParser<T extends QuerySpec> {
 
       return qp;
     }
-  }
-
-  public static Filter named(String name) {
-    return BY_NAME.get(name);
   }
 
 }
